+++ docs/CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to this project (rdpcloud) will be documented in this file.
 
 ## [dev] - DD/MM/YYYY
-- Updated docs
+### Changed:
+ - Updated the docs
+
+### Fixed:
+ - Windows domain / username check is case sensitive
 
 ## [0.1.0] - 19/10/2022
 - Initial Release